readyset-server: Basic node reuse at MIR generation
When adding a new node to the MIR graph, we check to see if an identical
node already exists.  If one does, we add the new query as an owner of
the node and return its index, throwing away the node we were
considering adding.

Release-Note-Core: Readyset will detect duplicate work between cached
  queries and eliminate it in order to speed propagation of writes
  through the system to the leaf caches.
Change-Id: Id8603a13fadd4debfee95b6a73bcaf26c6c24f35