readyset-sql-parsing: Parse custom `DROP` syntax atomically
Previously, if we parsed `ALL`, we could fail to parse either `CACHES`
or `PROXIED QUERIES`, and then succeed in parsing the `CACHE` branch as
if the `ALL` keyword was not present. This is incorrect, so we instead
check the entirety of the `ALL CACHES`/`ALL PROXIED QUERIES`/ branches
at once.

Change-Id: Ib05c911ae76ebf110d85d7dffd9a734571276f4d