dataflow-expression: Allow tuples in predicates
Allow queries like `SELECT * FROM t WHERE (a,b) IN ((1,2),(3,4))` and
`SELECT * FROM t WHERE (a,b) = (1,2)`.

Allow auto-parameterization and explicit parameterization of tuple
equality, but auto-parameterizing and explicit parameterization are not
yet supported with `IN`.

Note that tuples are implicit calls to the Postgres `ROW` constructor.
However, this doesn't mean that the `ROW` constructor (implicit or
explicit) is allowed in selects; this change only supports predicates.

Release-Note-Core: Readyset now allows tuple comparison in query
  predicates enabling syntax like
  `SELECT * FROM t WHERE (a,b) IN ((1,2),(3,4))`,
  `SELECT * FROM t WHERE (a,b) = (1,2)` and
  `SELECT * FROM t WHERE (a,b) = (?,?)`
Fixes: REA-5394
Change-Id: I79b46256d863796bf039ee212b53f6b9cf302b71
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8933
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>