readyset-sql-parsing: Add parity regression tests
These are basically just checking that the parser produces the same AST
as the old parser for some syntax that was either not hit by any
existing tests, or only hit by framework tests which we probably don't
want to rely on for this (or adding a test here just made fixing it
easier compared to relying on the framework test).

I've left in some ignored tests for future/in-progress functionality.

Change-Id: I21a0101cea23bc08db54961f0d61d22ffcfb6163
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8978
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>