readyset-sql: Lowercase unquoted function names in Postgres
The one downside here is there's no good way to borrow `ident` such that
it can be used in the error message in `next_expr()`, and I'd rather not
clone it just in case we need to use it. Replacing `next_expr` with a
macro would probably work, but I'm not convinced it's worth the effort.
If we actually see this error and can't figure it out, we can try that.

Change-Id: I076ae323c9d5965f8a166c3543eef12320cb789e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8994
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI