readyset-sql: Update to new sqlparser-rs
More index types have been added in sqlparser-rs. Rather than erroring
out or something for unsupported index types, lets just include them in
the AST. If they are useful one day, great, if not, 1) this is simpler
and less error prone, and 2) down the line we can piecemeal replace
parts of our AST with sqlparser's to remove the conversion overhead.

Change-Id: I71546e722819f59c3a5601273d4fabdfb2651b18