readyset-mysql: Set `?prefer_socket=false` when connecting to upstream
The default (`?prefer_socket=true`) can cause unexpected behavior if
running readyset on the same machine as a MySQL server that *isn't* the
intended upstream. With `?prefer_socket=true` (the default), Readyset
will reconnect to the intended MySQL server via TCP, but then attempt to
reconnect via the Unix socket as indicated in `@@socket`. If there is
actually a local MySQL with a socket with the same path, it will now be
connected to a server different from the one specified in the host and
port part of the connection string.

This patch will effectively hardcode `?prefer_socket=false` in a way
that can't be overridden (allowing overriding it would mean
re-implementing the `mysql` crate's URL parsing), so if you really want
to connect to the localhost MySQL server via the socket, you will have
to add `?socket=/path/to/socket` to your connection string.

Release-Note-Core: MySQL upstream connections no longer automatically
  attempt to reconnect via the Unix socket after a successful TCP
  connection on localhost. To connect to a local MySQL server via the
  socket, add `?socket=/path/to/socket` to the connection string.

Change-Id: I913b8d8883af25be59d60e8f7c1807ed574eeda2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9006
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI