sql-passes: Support IN op for tuples/rows in predicates
This is a follow-up to https://gerrit.readyset.name/c/readyset/+/8933
which added support for equality between tuples/rows in predicates.
This CL allows Readyset to rewrite queries with predicate
`(a,b) IN ((1,2),(3,4))` into a parametrized equality
`(a,b) = ($1, $2)` and query that multiple times with the input
parameters to produce the correct result.

Note that there is a hard-coded check to deal with the case where the
RHS list contains a single tuple. This is a limitation from nom-sql and
should be fixed by sqlparser.

Release-Note-Core: Readyset now allows parameterized IN operator
  between tuples in query predicates like
  `SELECT * FROM t WHERE (a,b) IN ((?,?),(?,?))`
  and auto-parameterizes queries like
  `SELECT * FROM t WHERE (a,b) IN ((1,2), (3,4))`.
Fixes: REA-5394

Change-Id: I5ad995e34651b72766e35b0ee4de2b6077890f3a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9026
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Tested-by: Buildkite CI