readyset-server: Check that placeholders' columns exist
Migration was not checking that the column connected to a placeholder
actually existed, and most confusingly, migration would even succeed!
We only hit a problem when querying the resultant cache, which wouldn't
work and would return an error that didn't make sense.  We now try to
check whether the column exists.

Release-Note-Core: Check whether placeholders' columns (`zzz = $1`)
  exist during migration, and if not, report query as unsupported.
Fixes: REA-5454
Change-Id: I33fa590bb51b19bc0355ebf13c042dcbebaf2a4c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9022
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>