replicators: Set MySQL replicator `character_set_results=utf8mb4`
We attempt to decode everything (including `SHOW CREATE TABLE` output)
as UTF-8. If the server defaults `character_set_results` to something
incompatible (e.g., `latin1`), we could easily fail and panic.

It is not ideal that the mysql library panics if it can't convert rows
instead of erroring, so we might look into fixing that somehow. (i.e.
`mysql::Conn::query` calls `from_row` instead of `from_row_opt`, and
this isn't configurable)

Change-Id: Ib784719c579a7022e7f58f488bc93accffd8632c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9034
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI