nom-sql: Lowercase unquoted no-paren functions in Postgres
This is just to add some consistency to our Postgres identifier casing
approach (which is to lowercase them during parsing), so that we can
treat identifiers uniformly when converting from sqlparser-rs. This
entire approach may go away in the future, and I would rather such a
hack/special case be in nom-sql rather than in readyset-sql.

Change-Id: I288e827d4ba3a3ca9bbc2552285777a187253d5a