Consider STRAIGHT_JOIN as a join operator
STRAIGHT_JOIN is a MySQL-specific join operator that is used to force
a join to be performed in the order of the tables specified. This is
done when the user suspects that the query planner is not choosing
the optimal join order.

This type of control is not supported by Readyset, so we don't need
to treat it differently from a regular JOIN/INNER JOIN.

We already parse STRAIGHT_JOIN as a join operator, but we don't
use it in the query graph. This commit fixes that.

Closes: REA-5502
Closes: #1468

Release-Note-Core: Add support for STRAIGHT_JOIN as a join operator.
Change-Id: Ibb145144373c32f5dbb1271bc0283fdca83a2659
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9044
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI