readyset-dataflow: apply default before coercion
When MySQL row based replication is enabled, if the type we are trying
to coerce to is a text with case-insensitive collation
(col_ty.is_citext()), we will attempt to coerce the value.
DfValue::Default is not coerceble, and should have been converted to
the column default value before attempting to coerce.

Closes: REA-5521
Closes: #1471

Release-Note-Core: Fixed a MySQL replication issue when Minimal Row
  Image is enabled. If a case-insensitive text column was omitted
  during an insert, we will try to coerce its value to the
  corresponding type, which will fail if we attempt to coerce before
  translating the value to the column default.

Change-Id: Ic95d3ad3ad53abcd61b13a06847b507841ef4705
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9077
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>