replicators: Add initial support for latin1 replication
Note that this only allows us to snapshot latin1 columns and read latin1
columns from the binlog. It simply causes the data to be decoded to utf8
before storage. This means that comparisons (cache lookup and ordering),
retrieval, and functions operating on this data will not match upstream.

This introduces a new `readyset_data::encoding::Encoding` type which
maps from MySQL collation IDs to corresponding encoders. We are using
`encoding_rs`, which implements the WHATWG Encoding Standard. This means
it does not precisely match MySQL's behavior for all encodings (e.g. it
uses a different replacement character than MySQL for invalid byte
sequences for some encodings). We may want to replace it in the future,
but it should be simple to swap it out within `readyset-data` to avoid
updating all the callsites (of which more will soon be added).

Release-Note-Core: As a first step toward supporting latin1 text columns
  in MySQL, we now allow replicating these columns but storing them
  re-encoded as utf8. Data will be retrieved as utf8, so comparisons and
  retrieval will not match upstream. Future work will improve support.

Change-Id: I86c4fbf8da2361ebbfa8ad6224bbce8763ad8e2d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9101
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI