readyset-data: Add encoding lookup for some missing supported collations
I missed these, including the default collation for utf8mb3, so that
test was only passing because we were defaulting to utf8.

But instead, we want better error messages for unknown/unsupported
encodings. Basically this just means we will log the MySQL collation ID
and fail instead of blindly attempting to decode using UTF-8.

Change-Id: I24b4b38e6690415424c0e71594ede7299ac6f0d5