encoding_mysql: Add test cases for different collations
This is basically just testing we are covering all the collation ids
actually sent by MySQL without simply hardcoding the list which would be
redundant, and with some added assurance that we are mapping to the
right encoding. But it might not be desirable for larger character sets
like utf8 ones which will take longer to test.

Change-Id: Iba5b556038bd05689ab3d0f77e1b53503cd115cf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9127
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>