readyset-sql: Use Vec instead of Option<Vec> for INSERT fields
MySQL is the only language that allows passing an empty list
of fields in an insert, but conceptually it's the exact same as
omitting the list. So it makes sense to have both be represented
with an empty vec instead of a None.

This is also how sqlparser does it and it's easier to change it
in RS rather in sqlparser.

Change-Id: I296e799e7983ededa7d9dfa93334d51d0a6aa0c0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9133
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI