readyset-server: Allow replicating with no fallback
The key change is that we were not starting replication if there was no
`--upstream-db-url`, but we can run perfectly fine without it and only
specifying `--cdc-db-url`.

Along with this, we cause most integration tests (those using
`TestBuilder`) which don't require fallback to run without it. Even
though we use `MigrationStyle::InRequestPath` for tests by default, we
will still proxy to upstream if the cache is dropped, for example. Then
a test might spuriously pass. Most tests don't want that, so we make
enabling fallback opt-in in the `TestBuilder`.

Change-Id: I60ac92ca6dabb2ff7ef0e780429ac20f6ab610dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9134
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>