replicators: Handle MySQL empty insert rows
MySQL, when configured with `binlog_row_image = minimal`, will only
send as part of an insert (`WRITE_ROWS_EVENT`), the columns that were
specified during the INSERT statement. The columns that are omitted
should receive the column default value.

There is an edge case when the user does not specify any columns,
meaning all the columns should receive it's default value. When this
happens, mysql sends an empty row. Our code will iterate over all rows
of the insert event via an iterator, and when there is an empty row,
our iterator is empty, which causes us to completely skip the insert.

Fixes: REA-5554
Closes: #1483

Release-Note-Core: Fixed a MySQL issue when Minimal Row-Based
  Replication is enabled and the user runs an insert with no columns.
Change-Id: I43fd6755f7568f0fbeef5dd84335129291116125
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9139
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>