readyset-mysql: Infer MySQL implicit defaults
MySQL allows users to omit columns from an INSERT statement, if the
column is defined as NOT NULL and has no default value it will
error out. However, if the user disable STRICT_TRANS_TABLES, MySQL
will infer the default value for the column.

This is fine for snapshot and replication with full row image, as we
will receive the inferred default value. However, this is not the case
for minimal row-based replication, where we receive only the columns
that were specified in the INSERT statement and are responsible for
inferring the default based on the column type, which we were
defaulting to None.

If a table has not PK or UK, DELETE/UPDATE operations will require
a full match on all columns to find the row to operate on. This will
cause replication to fail, as we and MySQL will have different values.

This commit also adjusts the mysql-time crate to not display a
negative value for zero time.

Closes: REA-5555
Closes: #1482

Release-Note-Core: Fixed a MySQL issue where INSERT statements with
  omitted columns could cause replication to fail under certain
  conditions.
Change-Id: Id00f4d6dc9b67368fad5453489f1768c4095ce7f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9140
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI