replicators: Refactor tests for better debuggability
It was more annoying to figure out where in the file errors were coming
from than it was to refactor this to just panic at the right spot in a
macro. Additionally, we now use `pretty_assertions` so that the errors
can be interpreted without hurting your eyes as much.

Change-Id: Ia8ef97ee23ae48705bf49b8f32667efa9a53d36e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9141
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>