readyset-server: Swap ControllerMessage and ReplicatorMessage names
Before, these names seemed exactly reversed to me, with a
ControllerMessage being processed by the replicator and a
ReplicatorMessage being processed by the controller.  The original
intent was for the name to signify who was sending the message.
However, these are multi-producer channels, and their identity need not
be tied to one particular sender.  Their usage has in fact already grown
beyond that.

Now, we swap the names such that a ControllerMessage is a message for
the controller to process and a ReplicatorMessage is a message for the
replicator to process.  Also standardize some variable names like
notification_channel into controller_tx, etc.

Change-Id: I0fc88858958dd5b98b6ee1b3ef62bee1d25170b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9148
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI