nom-sql: Fix compounded selects with ORDER BY and LIMIT
In a compounded select, the last ORDER BY and LIMIT clause
usually refer to the compound operator (EXCLUDE, UNION, etc..)
and not the last select (unless parenthesized).

While this shouldn't affect current behaviour of Readyset,
because of how we handle such queries, it will help in migrating
to sqlparser as it's flagged as a mismatch between nom and sqlparser.

Change-Id: I4797035baf0a84ff8a09941354e080a18f97a810
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9153
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>