encoding_mysql: Test replicating updates
The goal here is to ensure we get the same decoded data from the binlog
as we stored in RocksDB during snapshotting, to verify that our decoding
is matching the decoding that MySQL performs when we snapshot with `SET
character_set_results = 'utf8mb4'`.

Since this takes substantially longer than the existing testing, we
split the test into separate snapshot and streaming tests.

Change-Id: Ie75ecdb2a2b9732a0fb29862c240095049c3b5c1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9171
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>