sql-passes: Support aggregates inside other function calls
The pass responsible for breaking down query clauses previously
skipped `Expr::Call(f)` if `f` wasn’t an aggregate function
when listing the query’s aggregates.

This patch updates the pass to recurse into f if it’s not an
aggregate function. Currently, ROUND is the main example,
but this approach can be extended to support additional functions
as needed.

Fixes: REA-5422
Change-Id: I0015b2ebcf3676f112320565b3cbbcebfdec617b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9166
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI