readyset-data: Don't convert binary columns to text
We were automatically converting all incoming binary columns (e.g.
`BLOB`, `BINARY`) to `DfValue::Text` iff it contained valid UTF-8. This
led to different behavior between rows for the same column, and changed
the byte-level value when returned if it did have valid UTF-8 but the
representation didn't match. This was additionally not matching MySQL
because the resulting `DfValue::ByteArray`s were being incorrectly
converted to their hex representation before being returned.

Unfortunately, we were relying on this implicit conversion to text for
several things including converting `ENUM` and `NUMERIC` values (which
arrive as bytes) to `DfValue::Integer` and `DfValue::Numeric` by
matching/parsing the string we get by first turning it into
`DfValue::Text`. To simplify this patch, we simply replicate that
behavior by implementing `DfValue::coerce_to` for `DfValue::ByteArray`
by first attempting to coerce it to `DfValue::Text` (as we would have
done before) and the coercing that.

Change-Id: Id90a901e4fde5b64de099af39a946db249c3b5c1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9188
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>