readyset-server: Create a shareable pool of controller connections
Previously, each time a new controller was elected, we would create for
the worker a single connection to that controller.

Now, we create a pool of connections shared between the controller and
worker, which is behind a read-write lock.  When a new controller is
elected as leader, we take a write lock and replace the existing pool of
connections with a pool of connections pointing at the new address.

When a connection is desired from the pool, we take a read lock and then
ask the pool for a connection.

The idea here is to allow for passing a reference to this pool of
controller connections around to various places in readyset-server, such
that they can then easily make requests to the controller.

This allows us to do a little clean up as well.  The controller no
longer needs the worker_tx channel or the ability to send worker
requests, as these were only used to inform the worker of a leader
change.  Similarly, we can get rid of WorkerRequestKind::NewController,
as this message was only used for this purpose.

Addresses: REA-2463, REA-2907, REA-4863, REA-5751
Change-Id: Ied736cd0c4935849f860281d289e21c7ad314cfb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10195
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI