MySQL Minimal Row Base Replication: Fix collation and signedness
MySQL MRBR differs from full row as it only sends the columns that
were inserted/updated. The table map event contains the metadata for
all the columns, not only the ones present in the event, as in case of
multiple rows inserted, write/update events might share a single TME.

One special part of the TME is the optional metadata types[1], which
contains information such as collation and signedness of the column.
The way this is exposed by our rust mysql driver is via one iterator,
which gets populated only for columns that are relevant for that
subject. For example, for the below table:

```sql
CREATE TABLE `b` (
  `col_1` char(1) collate utf8mb4_0900_ai_ci,
  `col_2` int,
  `col_3` unsigned int,
  `col_4` char(1) collate binary
);
```

The iterators will be:
```
charset_iter: [255, 63]
signedness_iter: [true, false]
```

When we do an UPDATE b SET col4 = 'A' the row event will come as:

```
{@3: Value(Bytes("A"))}
```

Making it difficult to retrieve the charset 63 for col4.

Our current code walks the columns from the row event, and depending
on the column type, it advances either the charset or the signedness
iterator. This causes the above update to retrieve charset 255 for
col4.

This commit fixes binlog_row_to_noria_row to create a vector with
charset and signedness for each column, respecting the index received
in the row event, making it possible to retrieve the correct values
even when columns are skipped.

[1]: https://dev.mysql.com/doc/dev/mysql-server/latest/classmysql_1_1binlog_1_1event_1_1Table__map__event.html

fixes: REA-5634
fixes: #1491

Release-Note-Core: Fixes an issue with MySQL Minimal Row Based
  Replication where the collation and signedness of columns were
  incorrectly retrieved from the table map event.
Change-Id: I422408f60421b53354e076c1599ebacba9e8b89a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9219
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI