readyset-data: Add Utf8AiCi collation
On MySQL we previously interpreted any collation ending with "_ci" as
being our own custom thing that loosely attempted to be case
insensitive, but we made no attempt at either accent insensitivity (the
default in MySQL) or normalization (e.g. treating characters with
combining accents as equal to pre-made counterparts).  Now we push that
stuff over to icu4x.

Change-Id: I1798dec130f3a75dd51df10e88c052e312ead261
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9234
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI