readyset-data: Make collation selection interface more generic
Rather than having a single function that is specific to MySQL, we now
accept a dialect to aid our interpretation of collation names.  Some of
these helpers could become putlic if they seem useful in other crates.

Postgres seems to have no collation selection support at all, simply
relying on the Default implementation, but that's a battle for another
day.

Change-Id: I93c1419e9ca486c6489aa58f2ed0125a4e8d3907
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9235
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>