readyset-adapter: Drop cache before creating its replacement
When creating a named cache in Readyset, we first drop any existing
caches with that name.  For unnamed caches, we did something different.

Previously, if you created an unnamed cache, we would implicitly give it
a name based on its query id.  If a duplicate cache already existed with
that query id, we would attempt to reuse that existing cache, but not
before we would start the process of creating a new dataflow graph, but
then ultimately bail out instead of completing it.

Unfortunately, marching forward like this caused an issue that left our
graph metadata in SqlToMirConverter in a state where some stored node
index values were updated to new values, even though we never actually
updated the graph itself to use these new values.  Due to these
undesired changes to node index metadata, future attempts to drop and
recreate the cache could fail, as we might attempt to remove a node
index that didn't actually correspond with the desired node in the
graph.

Now, we first drop such caches and recreate them.  This is the same
thing we do with named caches.

There is still room for improvement in the sense that we could try to
avoid starting the graph creation process at all, and also a partially
executed dataflow compilation should not leave some metadata in a bad
state, but these goals seem like larger projects.

Fixes: REA-5616, REA-5636
Change-Id: Iad22f58c545d652caaac8506cb4bdb23955c8d5a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9271
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI