benchmarks: Fix conversion of results to text
We were relying on the implicit conversion of any `&[u8]` that happens
to be a valid UTF-8 string into `DfValue::Text`. This is a bit of a hack
in the benchmarking code where we are using conversions to `DfValue` to
paper over the possible differences between a Postgres and MySQL
upstream. So I'm not too worried about this fix being hackish too.

Change-Id: Ibdd1a5bb119cba5d8b419677b14af511b8d7d701
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9272
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI