benchmarks: Initialize tracing logging
We previously weren't getting any logging output from the benchmark
workload side of things because we never initialized a tracing
subscriber. Unfortunately, the easiest way to do it is to initialize it
after forking, because otherwise the adapter that's spawned in the child
process will try to set up its own tracing subscriber and puke if it
already finds one.

Change-Id: I126d3c5ddf83d87181feb1ab55a65abab8d7d701
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9273
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>