Remove mysql-simple dependency
We had two tests that required mysql-simple dependency. This caused
a maintanance burden, as mysql_async and mysql-simple are released
with different cadence, and we need to ensure we were using a version
that linked to the same version of mysql_common.

MySQL simple was used on two tests:
1) mysql-srv/main test - Fixed the tests to use mysql_async.
2) mysql_comparison - This file boots a server and run one single test.
The test is disabled for about 7 years already and looking into the
blame into the file, looks like we are often mantaining the code just
for the sake of compiling it. I removed test completelly.

Change-Id: Ic2a3995fbdb5e212a95bd22f28e24b1fb06fdc42
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9282
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>