readyset-mysql: Track MySQL character sets seen in new metric
When we encounter a `SET NAMES <charset>` or `SET character_set_results
= <charset>` statement, we now increment a counter to track that we've
seen that character set. We also label which statement type saw it.

Note that this doesn't account for upstream defaults, which will require
a different CL which updates `UpstreamSystemProperties`.

We partially plumb the character set in the form of an encoding, but it
is not yet utilized by the connector or backend.

Fixes: REA-5709

Change-Id: Ic6042cd0b5eb2d33406415957c46bec0ea83ebf4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9346
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>