readyset-data: Add `DfType::Point`
Add a new `DfType` to support an initial spatial data type, which is
`Point`. This CL also adds basic MySQL support in
replicator (snapshot, replication).

A logictest is provided to test the output is correct (matches the
upstream's output). It is expected, though, that real-world queries
will use one of the `ST_*` spatial functions to select the data; those
functions are inplemented in a follow-up CL.

Fixes: REA-5697
Release-Core-Note: Basic support for `Point` spatial type in mysql.
Change-Id: If79e5385ec32d9f58be601c0366312b4ddfad6a6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9320
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI