dataflow-expression: Spatial display function support
Add support for spatial "display" functions `ST_AsText` and `ST_AsWKT`
for mysql `Point` data types. In MySQL, the two functions are the
same (one is an alias of the other).

The logictest for point types is extended to test these new functions.

Fixes: REA-5697
Release-Core-Note: Add support for `ST_AsText` and `ST_AsWKT`
  functions.
Change-Id: I8796c7effb99b7e85b6d66f82df4f1e24d9f792b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9322
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI