Fix nightly failure in fallback tests
We changed the behavior of CREATE CACHE in Iad22f58c such that we now
first drop a duplicate, implicitly-named cache before continuing with
with the creation process.

Due to this change, a couple tests in nightly began to fail, as they
expected a reason for why we would upstream a query, which is now
different due to the new recreation behavior.

This patch reworks the tests to expect the new behavior of a duplicate
cache creation.  To make the testing of expected upstream reason more
symmetric, we now also test that cache creation for 2 different queries
fail in the expected manner, as opposed to just 1 of them.

Also improve some error reporting in the tests if the values are not
what we expect.

Addresses: REA-5616, REA-5636
Change-Id: I360bdea59eb374ea5a99151cafaacbaecac67229
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9313
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>