build: Don't use the special nightly mirror for the readyset-build image
This is pointless. Per the [description] of the image: "The nightly tag is
configured identically to that of the latest tag of the official image,
except that the nightly toolchain is selected via rustup."

Guess what is the very first thing we do when building this image?
Select a different toolchain via rustup.

[description]: https://hub.docker.com/r/rustlang/rust

Change-Id: Ia6dff6e633c9a650f53ff3fedaa96e0acdfb4b0f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9355
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Reviewed-by: Ron Hough <ron@readyset.io>