Reject NUMERIC/DECIMAL values with scale > 27
Previously, we enforced a scale limit of 28 due to a limitation in
rust_decimal.  We were able to enforce this limit by detecting scales
larger than 28 in its returned Decimal struct.

Unfortunately, our mechanism for detecting invalid Decimal struct values
has been compromised by a recent change to rust_decimal, which now
rounds PostgreSQL NUMERIC scales above 28 down to 28.  This means that a
Decimal provided by the library with a scale of 28 could have actually
originated from a number with a larger scale, which has been erroneously
truncated to 28.

Now, we reduce our allowable scale to 27 and reject scale values of 28
or higher, as they may be incorrect.

Fixes: REA-5719
Change-Id: I7a041702cce0e55132145afb83901f8914ec618d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9372
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>