replicators: correctly handle PostgreSQL table snapshot errors
If more tables are queued for snapshot than the configured
max_parallel_snapshot_tables, and at least the first table to finish
fails, the failure could be silently ignored.

This happened because the wait loop only monitored the number of
in-progress snapshots, without verifying if any completed snapshot
had failed. As a result, we continued snapshotting subsequent tables
without properly reporting the failure. This caused the failed table
to still be marked as in-progress (having an entry in the offsets but
with a None value), which we caught by the panic in the end of the
snapshot loop.

Fixes: REA-5722
Closes: #1510

Release-Note-Core: Fixed a bug where a PotgreSQL table snapshot
  failure could be silently ignored, causing Readyset to fail to
  remove the failed table from the list of tables to snapshot, and
  abort the entire snapshot process.
Change-Id: Ib1dcabbb11531fb515d71777124f4bec08774965
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9374
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>