vertical: Create reqwest::Client once
A bisect of the bulk package update in I20155667 revealed that upgrading
the native-tls package to version 0.2.14 introduced a performance
regression in our nightly run of readyset-mysql/tests/vertical.rs
(actually, the problem starts with version 0.2.13).

This issue has been identified upstream and a fix was attempted here:
https://github.com/sfackler/rust-native-tls/issues/320

The gist is that version 0.2.12 used to perform
openssl_probe::init_ssl_cert_env_vars() once during initialization, and
version 0.2.13 moved away from this interface in favor of
openssl_probe::probe().  Although version 0.2.14 attempts to improve
performance by caching the call to openssl_probe::probe(), our
situtation does not seem to improve.

Our test created a reqwest::Client for every run of Operation::Evict,
which exacerbated the issue, leading to slower runs.  Let's adjust our
vertical tests such that they only create a single reqwest::Client,
sharing it in the tests.  Creating one reqwest::Client like this shaves
about 20 minutes off of its duration prior to the regression we saw in
I20155667.

This regression was not noticed on MacOS, as the native-tls package uses
the security-framework package as opposed to the openssl package when
running in that environment.

Change-Id: Ia1a8687be7c3756debc7f461f612f9d0c51b8339
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9409
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI