replicators: Capture custom postgres OID name
Currently when we receive a WAL message (in the postgres logical
replication) that defines a custom data type, we only keep the OID
around in a `HashSet`. This CL changes the `HashSet` to ` HashMap` so
we can retain the `name` of the type, as well.

The motivator for this is that when operators use a postgres
extension, like `pgvector` or `postgis`, the extension may install
custom data types, but the OIDs are not guaranteed to be consistent
across different postgres installation. Thus, for example, we cannot
rely on OID "12345" to represent a postgis `geometry(point)` across
every postgres database. We can, however, make use of the data type's
name for that uniqueness.

Change-Id: If4f76b00cdbcfe3025489127e5226ec6c9b2f6ad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9419
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI