replicators: Support non-utf8 character sets in MySQL 5.7
MySQL 5.7 does not send column character set information with table map
events (not even as a `DEFAULT_CHARSET` optional metadata TLV, because
it doesn't even have the optional metadata field). You are expected to
rely on previous knowledge of the DDL for the given table.

We now fall back to looking up the character set from the table's schema
as it is stored in the controller. This introduces a round trip to the
controller, so we store the table schemas we have retrieved.

Fixes: REA-5725

Change-Id: I01467bda18613859debde26d44f5fe51e740d410
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9432
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI