treewide: Add `mysql8_upstream` test tag
This is for tests that can't or don't currently work against MySQL 5.7.

Currently these are due to:

- Missing access-sensitive/insensitive utf8 collations.
- Different timestamp behavior that I haven't looked into yet.
- No support for `@@binlog_transaction_compression`.
- Lower default `@@max_allowed_packet`.
- Some different `MEDIUMINT` and `BIGINT` behavior I haven't looked into
  yet.
- Different expected column definition flags I haven't looked into yet.
- Missing `json_overlaps` function.

Several of these could probably be adapted to work with 5.7, e.g.
setting `@@max_allowed_packet` either in the docker compose or at the
start of that test. Some like accent sensitivity and
`@@binlog_transaction_compression` simply don't exist on 5.7.

Change-Id: I2b1099fcdf0a5567caf1b0e7948694246bb986c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9437
Reviewed-by: Ron Hough <ron@readyset.io>
Tested-by: Buildkite CI