readyset-server: Bump soft file descriptor limit to 1048576 on Linux
There's not really a good reason to limit ourselves here.  Let's
try to increase our soft file descriptor limit to the default max value
allowed for a process by the Linux kernel:
https://docs.kernel.org/admin-guide/sysctl/fs.html#nr-open

It's possible that we'll hit the hard limit before this.  Log a line at
INFO if we do.

Change-Id: Ib4b8eda08d239541f0734325605112bacc8f74b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9445
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>