treewide: Replace rust_decimal with bigdecimal
We wrap the type with our own enum capable of representing +/-Infinity
and NaN. This allows us to snapshot and replicate tables with any
`DECIMAL`/`NUMERIC` precision, removing the limit of 27 (previously 28)
digits.

`bigdecimal` is certainly slower for very large numbers, and it may be
worthwhile to add a `Decimal::SmallNumber` variant which can fit into
`rust_decimal`'s i128 representation, but I don't know how much faster
it is for smaller numbers.

Fixes: REA-1933, REA-5720
Release-Note-Core: Add support for arbitrarily large `DECIMAL` and
  `NUMERIC` values. We can now replicate and use the full range allowed
  by MySQL and Postgres for these columns. This also fixes an issue with
  replicating NaN values, which will now work as expected.
BuildDistroReleasePkg: decimalalpha

Change-Id: I9ed0219e1f78d5e1aa8466a41ff6d42239fb9467
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9519
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI